Move GOV.UK Notify related functions to separate module
